+++ CHANGELOG.next.toml
@@ -38,3 +38,21 @@ message = "Fix bug where ECS credential provider could not perform retries."
 meta = { "breaking" = false, "tada" = false, "bug" = true }
 references = ["smithy-rs#998", "aws-sdk-rust#359"]
 author = "rcoh"
+
+[[aws-sdk-rust]]
+message = "`aws_http::AwsErrorRetryPolicy` was moved to `aws_http::retry::AwsErrorRetryPolicy`."
+meta = { "breaking" = true, "tada" = false, "bug" = false }
+references = ["smithy-rs#todo"]
+author = "rcoh"
+
+[[smithy-rs]]
+message = "The signature of `aws_smithy_protocol_test::validate_headers` was made more flexible but may require adjusting invocations slightly."
+meta = { "breaking" = true, "tada" = false, "bug" = false }
+references = ["smithy-rs#todo"]
+author = "rcoh"
+
+[[aws-sdk-rust]]
+message = "Add recursion detection middleware to the default stack"
+meta = { "breaking" = false, "tada" = false, "bug" = false }
+references = ["smithy-rs#todo"]
+author = "rcoh"
