+++ codegen/src/main/kotlin/software/amazon/smithy/rust/codegen/smithy/generators/protocol/ProtocolTestGenerator.kt
@@ -366,7 +366,7 @@ class ProtocolTestGenerator(
             return
         }
         val variableName = "expected_headers"
-        rustWriter.withBlock("let $variableName = &[", "];") {
+        rustWriter.withBlock("let $variableName = [", "];") {
             write(
                 headers.entries.joinToString(",") {
                     "(${it.key.dq()}, ${it.value.dq()})"
